requirements.txt
fastapi
uvicorn[standard]
boto3
jinja2
python-multipart
python-dotenv

